♻️ REFACTOR: 어드민 공고 등록 폴더 규칙에 알맞게 코드 정리 #143
