+++ src/entities/admin-review-detail/model/admin-review-detail.types.ts
@@ -1,0 +1,14 @@
+import {
+  AdminAdditionalOnboardingSchema,
+  AdminAnnouncementRequestSchema,
+  AdminAnnouncementSchema,
+} from "@/src/entities/admin-review-detail";
+import { z } from "zod";
+
+export type AdminAnnouncementResponse = z.infer<typeof AdminAnnouncementSchema>;
+export type AdminAdditionalOnboardingResponse = z.infer<
+  typeof AdminAdditionalOnboardingSchema
+>;
+export type AdminAnnouncementRequest = z.infer<
+  typeof AdminAnnouncementRequestSchema
+>;
